pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "er-triage-env"
version = "1.0.0"
description = "OpenEnv ER Triage Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.0",
    "fastapi",
    "uvicorn",
    "openai"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server", "client"]
py-modules = ["models"]
